fix: fixed imports; moved Seperator to correct directory
